pyproject.toml
[project]
name = "mirror-mirror"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.7"
dependencies = [
    "accelerate>=1.7.0",
    "betterproto>=2.0.0b7",
    "diffusers>=0.33.1",
    "faststream[cli,redis]>=0.5.40",
    "google-genai>=1.15.0",
    "huggingface-hub[hf-xet]>=0.31.2",
    "janus>=2.0.0",
    "nvidia-cublas-cu12>=12.6.4.1 ; sys_platform == 'linux'",
    "nvidia-cuda-cupti-cu12>=12.6.80 ; sys_platform == 'linux'",
    "nvidia-cuda-nvrtc-cu12>=12.6.77 ; sys_platform == 'linux'",
    "nvidia-cuda-runtime-cu12>=12.6.77 ; sys_platform == 'linux'",
    "nvidia-cudnn-cu12>=9.5.1.17 ; sys_platform == 'linux'",
    "nvidia-cufft-cu12>=11.3.0.4 ; sys_platform == 'linux'",
    "nvidia-cufile-cu12>=1.11.1.6 ; sys_platform == 'linux'",
    "nvidia-curand-cu12>=10.3.7.77 ; sys_platform == 'linux'",
    "nvidia-cusolver-cu12>=11.7.1.2 ; sys_platform == 'linux'",
    "nvidia-cusparse-cu12>=12.5.4.2 ; sys_platform == 'linux'",
    "nvidia-cusparselt-cu12>=0.6.3 ; sys_platform == 'linux'",
    "nvidia-nccl-cu12>=2.26.2 ; sys_platform == 'linux'",
    "nvidia-nvjitlink-cu12>=12.6.85 ; sys_platform == 'linux'",
    "nvidia-nvtx-cu12>=12.6.77 ; sys_platform == 'linux'",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "psutil>=7.0.0",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "redis>=6.0.0",
    "rich>=14.0.0",
    "tenacity>=9.1.2",
    "torch>=2.5.0",
    "transformers>=4.51.3",
    "numpy < 2",
    "triton>=3.3.0 ; sys_platform == 'linux'",
    "typer>=0.15.3",
    "pygame>=2.6.1",
]

[project.scripts]
diffusion-server = "mirror_mirror.diffusion_server:app"
pygame-display = "pygame_display_subscriber:app"

[build-system]
requires = ["uv_build>=0.7.3,<0.8.0"]
build-backend = "uv_build"

[tool.uv.sources]
betterproto = { git = "https://github.com/danielgtaylor/python-betterproto", rev = "124613f55fa8ff62a460e1ad33c5c78723b703a0" }
torch = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { marker = "sys_platform == 'linux'", path = "./wheels/torch-2.5.0a0+872d972e41.nv24.08-cp310-cp310-linux_aarch64.whl" },
]
torchvision = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  # { index = "pytorch-cu", marker = "sys_platform == 'linux'" },
]
triton = [
  { marker = "sys_platform == 'linux' and platform_machine == 'aarch64'", url = "https://pypi.jetson-ai-lab.dev/jp6/cu126/+f/7e8/bb11e12039b80/triton-3.3.0-cp310-cp310-linux_aarch64.whl"},
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.ruff]
line-length = 120

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
    "pytest-mock>=3.14.1",
    "pytest-timeout>=2.4.0",
]
